requirements.txt
invoke==2.2.0
torch==2.5.1
typer==0.15.1
wandb==0.19.1
matplotlib==3.10.0
hydra-core==1.3.2

